Add HACK? for issues with Bundler
Even with LD_LIBRARY_PATH present the graphql-libgraphqlparser (or
Bundler) cannot seem to find the library. As a work around we export the
path where the library and headers can be found.

Compounding this:

Bash does not allow us to export variables that include dashes in their
name. Unfortunately this means that we cannot export the required
environment variable that Bundler expects since the Gem has a dash in
it's name.

To work around this we conditionally create a .bundle/config file and
add the build configuration for the gem.